requirements.txt
Flask==3.0.3
python-docx==1.1.2
requests==2.32.3
Werkzeug==3.0.3
gunicorn